Do not send 'Origin' header in handshake
This is a header only intended to support for Web Origin security in
browsers. Other clients are not required to include this header, and the
current case complicates servers since it receiving 'ws://' or 'ws://'
in the 'Origin' header is highly unusual.

RFC 6455, Section 4.1, Client Requirements:

> Additionally, if the client is a web browser, it supplies /origin/.

RFC 6455, Section 4.2.1, Reading the Client's Opening Handshake:

> Optionally, an |Origin| header field.  This header field is sent by
> all browser clients.  A connection attempt lacking this header field
> SHOULD NOT be interpreted as coming from a browser client.